Initialize the specstesttool branch, which aims to resturcture specstest as a tool for specstesting.
