--- cases/specstest/linux_capabilites/linux_capabilities_SETFCAP_runc.json
@@ -1,63 +1,0 @@
-{
-  "Name": "Specstest-linux-capabilities-SETFCAP",
-  "Summary": "test linux-capabilities-SETFCAP",
-  "Owner": "wangqilin2@huawei.com",
-  "Description": "Test runc  whether it supports the linux capabilities SETFCAP",
-  "Group": "Specstest/linux_capabilities/",
-  "License": "Apache 2.0",
-  "Explains": "Test runc by set the capabilities SETFCAP to an other file",
-  "Requires": [
-    {
-      "Class": "OperationOS",
-      "Type": "os",
-      "Distribution": "ubuntu",
-      "Version": "14.04",
-      "Resource": {
-        "CPU": 1,
-        "Memory": "1GB",
-        "Disk": "2G"
-      }
-    },
-     {
-      "Class": "specstest",
-      "Type": "container",
-      "Distribution": "runc",
-      "Version": "0.2",
-      "Files": [
-        "./source/config.json"
-      ]
-    }
-  ],
-  "Deploys": [
-    {
-      "Object": "hostA",
-      "Class": "OperationOS",
-      "Cmd": "go build linux_capabilities_SETFCAP_host.go ; ./linux_capabilities_SETFCAP_host;",
-       "Files": [
-        "./source/linux_capabilities_SETFCAP_guest.go",
-        "./source/linux_capabilities_SETFCAP_host.go"
-      ],
-      "Containers": [
-        {
-          "Object": "specs",
-          "Class": "specstest"
-        }
-      ]
-    }
-  ],
-  "Run": [
-    {
-      "Object": "hostA",
-      "Class": "OperationOS",
-      "Cmd": " cd  ./../../source/;runc"
-    }
-  ],
-  "Collects": [
-    {
-      "Object": "hostA",
-      "Files": [
-        "/tmp/testtool/linux_capabilities_SETFCAP.json"
-      ]
-    }
-  ]
-}
