--- cases/specstest/linux_resources/linux_resources_memory_limited.json
@@ -1,55 +1,0 @@
-{
-  "Name": "Specstest-linux-resources-memory-limited",
-  "Summary": "test linux_resources_memory_limited",
-  "Owner": "wangqilin2@huawei.com",
-  "Description": "Test runc  whether it supports the linux memory size limited",
-  "Group": "Specstest/linux_resources/",
-  "License": "Apache 2.0",
-  "Explains": "Test runc by see whether it can constrain the  cgroup memory size",
-  "Requires": [
-    {
-      "Class": "OperationOS",
-      "Type": "os",
-      "Distribution": "Ubuntu",
-      "Version": "14.04",
-      "Resource": {
-        "CPU": 1,
-        "Memory": "1GB",
-        "Disk": "1G"
-      }
-    },
-    {
-      "Class": "specstest",
-      "Type": "container",
-      "Version": "runc V0.2"
-    }
-  ],
-  "Deploys": [
-    {
-      "Object": "hostA",
-      "Class": "OperationOS",
-      "Cmd": "go build linux_resources_memory_limited_host.go ; ./linux_resources_memory_limited_host; go build linux_resources_memory_limited_runcrunner.go; cp ./linux_resources_memory_limited_runcrunner ./../../source/;",
-      "Containers": [
-        {
-          "Object": "specs",
-          "Class": "specstest"
-        }
-      ]
-    }
-  ],
-  "Run": [
-    {
-      "Object": "hostA",
-      "Class": "OperationOS",
-      "Cmd": "cd  ./../../source/;./linux_resources_memory_limited_runcrunner"
-    }
-  ],
-  "Collects": [
-    {
-      "Object": "hostA",
-      "Files": [
-        "/tmp/testtool/linux_resources_memory_limited.json"
-      ]
-    }
-  ]
-}
