--- cases/specstest/linux_resources/source/linux_resources_memory_limited_guest.go
@@ -1,35 +1,0 @@
-// Copyright 2014 Google Inc. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package main
-
-import (
-	"log"
-	"os/exec"
-)
-
-func resourcesMemoryLimited() {
-	for i := 1; i >= 0; {
-		cmd := exec.Command("/bin/sh", "-c", "echo \"runc container keep running\" ")
-		_, err := cmd.Output()
-		if err != nil {
-			log.Fatalf("[Specstest] linux resources memory limited test:  runc running error, %v", err)
-		}
-	}
-
-}
-
-func main() {
-	resourcesMemoryLimited()
-}
