--- cases/specstest/mount_fstype/mount_fstype_support_tmpfs.json
@@ -1,59 +1,0 @@
-{
-  "Name": "Specstest-mount_fstype_tmpfs",
-  "Summary": "test based opencontainers/specs",
-  "Owner": "wangqilin2@huawei.com",
-  "Description": "Test runc whether it can mount the tmpfs filesystem",
-  "Group": "Specstest/mount_fstype/",
-  "License": "Apache 2.0",
-  "Explains": "Test whether the container supports to mount a filesystem",
-  "Requires": [
-    {
-      "Class": "OperationOS",
-      "Type": "os",
-      "Distribution": "ubuntu",
-      "Version": "14.04",
-      "Resource": {
-        "CPU": 1,
-        "Memory": "1GB",
-        "Disk": "2G"
-      }
-    },
-    {
-      "Class": "specstest",
-      "Type": "container",
-      "Distribution": "runc",
-      "Version": "0.2",
-      "Files": [
-        "./source/config.json"
-      ]
-    }
-  ],
-  "Deploys": [
-    {
-      "Object": "hostA",
-      "Class": "OperationOS",
-      "Cmd": "go build mount_fstype_host.go;./mount_fstype_host --fs tmpfs ; cp mount_fstype_runcrunner.sh ./../../source/ ",
-      "Containers": [
-        {
-          "Object": "specs",
-          "Class": "specstest"
-        }
-      ]
-    }
-  ],
-    "Run": [
-    {
-      "Object": "hostA",
-      "Class": "OperationOS",
-      "Cmd": " cd  ./../../source/;./mount_fstype_runcrunner.sh tmpfs"
-    }
-  ],
-  "Collects": [
-    {
-      "Object": "hostA",
-      "Files": [
-        "/tmp/testtool/mount_fstype_out.json"
-      ]
-    }
-  ]
-}
