--- cases/specstest/mount_fstype/source/mount_fstype_runc.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-cd ./../../source ; runc 
-if [ $? -eq 0 ]; then
-	echo "[NO]   Filesystem $1 Mount Unsupported "
-else 
-	echo "[Yes]   Filesystem $1 Mount Supported "
-fi
-
-
