--- cases/specstest/namespace/source/uts.go
@@ -1,49 +1,0 @@
-// Copyright 2015 Huawei Inc. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package main
-
-//import "github.com/opencontainers/specs"
-//shoud use github path ,but the whole project is not supported
-import "../../source/specs"
-
-func setUtsSpec(ns specs.Namespace) specs.LinuxSpec {
-	spec := linuxSpec
-	spec.Linux.Namespaces = append(spec.Linux.Namespaces, ns)
-	spec.Process.Args = append(spec.Process.Args, "/bin/readlink", "/proc/self/ns/uts")
-	return spec
-
-}
-
-func TestUtsPathEmpty() string {
-
-	ns := specs.Namespace{Type: "uts",
-		Path: ""}
-
-	ls := setUtsSpec(ns)
-	result, err := TestPathEmpty(&ls, "/proc/*/ns/uts")
-
-	return MarshalTestResult("TestUtsPathEmpty", ns, err, result)
-}
-
-func TestUtsPathUnempty() string {
-
-	ns := specs.Namespace{Type: "uts",
-		Path: "/proc/1/ns/uts"}
-
-	ls := setUtsSpec(ns)
-	result, err := TestPathUnEmpty(&ls, ns.Path)
-	return MarshalTestResult("TestUtsPathUnempty", ns, err, result)
-
-}
