--- cases/specstest/namespace/specstest-namespace.json
@@ -1,65 +1,0 @@
-{
-  "Name": "Specstest-namespace-runc",
-  "Summary": "test based opencontainers/specs",
-  "Owner": "liugenping@huawei.com",
-  "Description": "Test runc's namespace",
-  "Group": "Specstest/namespace/",
-  "License": "Apache 2.0",
-  "Explains": "",
-  "Requires": [
-    {
-      "Class": "OperationOS",
-      "Type": "os",
-      "Distribution": "Ubuntu",
-      "Versioin": "14.04",
-      "Resource": {
-        "CPU": 1,
-        "Memory": "1GB",
-        "Disk": "2G"
-      }
-    },
-    {
-      "Class": "specstest",
-      "Type": "container",
-      "Distribution": "runc",
-      "Version": "0.2"
-    }
-  ],
-  "Deploys": [
-    {
-      "Object": "hostA",
-      "Class": "OperationOS",
-      "Files": [
-        "../../source/config.json"
-      ],
-      "Cmd": "go build namespace.go  ipc.go  network.go pid.go testframe.go uts.go mount.go",
-      "Containers": [
-        {
-          "Object": "specs",
-          "Class": "specstest"
-        }
-      ]
-    }
-  ],
-  "Run": [
-    {
-      "Object": "hostA",
-      "Class": "OperationOS",
-      "Cmd": "./namespace",
-      "Containers": [
-        {
-          "Object": "specs",
-          "Class": "specstest"
-        }
-      ]
-    }
-  ],
-  "Collects": [
-    {
-      "Object": "hostA",
-      "Files": [
-        "./source/namespace_out.json"
-      ]
-    }
-  ]
-}
