--- cases/specstest/platform_linux_amd64/source/demo.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"fmt"
-)
-
-func main() {
-	fmt.Println("runc ran!")
-}
