--- cases/specstest/report/README.md
@@ -1,64 +1,0 @@
-Specstest based on the specsversion: commmit  7414f4d3e90b5c22cae7c952d123e911c0cf94ba on  [opencontainers/specs](https://github.com/opencontainers/specs)
-Every item of the first level config have one testcase now.
-
-### Next to Do
-- Rich testcases to cover all of the config item on the specsversion: commmit  7414f4d3e90b5c22cae7c952d123e911c0cf94ba.
-- Do test on the newest specsversion, output which config item not supported yet when not supported well by runc .
-- Framework restructure to support the auto specsversion specified.
-- Specstest on other bundles/containers.
-
-### What the TestResult Mean
-for example, in process_out.json
-``` json
-{
-  "Linuxspec.Spec.Process": {
-    "terminal": {
-      "true": "passed"
-    },
-    "user": {
-      "uid": {
-        "1": "passed"
-      },
-      "gid": {},
-      "additionalGids": {
-        "nil": "passed"
-      }
-    },
-    "args": {
-      "./process_guest": "passed"
-    },
-    "env": {
-      "PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin": "passed",
-      "TERM=xterm": "passed"
-    },
-    "cwd": {
-      "/testtool": "passed"
-    }
-  }
-}
-```
-
-It means, when Linuxspec.Spec.Process 's config is as below, the testresult is passed, the specs item is supported by runc.
-``` json
- "process": {
-    "terminal": "true",
-    "user": {
-      "uid": 1,
-      "gid": 1,
-      "additionalGids": null
-    }
-```
-Anyway, in order to test process config, the project need other config item support, it is the input of the testcase, so the testsesult also give it out,
-``` json
-  "args": {
-      "./process_guest": "passed"
-    },
-    "env": {
-      "PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin": "passed",
-      "TERM=xterm": "passed"
-    },
-    "cwd": {
-      "/testtool": "passed"
-    }
-  }
-```
