--- cases/specstest/report/linux_capabilities_SETFCAP.json
@@ -1,5 +1,0 @@
-{
-  "Linuxspec.Linux.Capabilities": {
-    "SETFCAP": "pass"
-  }
-}
