--- cases/specstest/report/linux_resources_memory_limited.json
@@ -1,1 +1,0 @@
-{"Linuxspec.Linux.Resources.Memory":{"Memory.Limit":"pass"}}
