--- cases/specstest/report/mount_fstype_out.json
@@ -1,5 +1,0 @@
-{
-  "Linuxspec.Mounts.Filesystem": {
-    "tmpfs": "success"
-  }
-}
