--- cases/specstest/report/namespace_out.json
@@ -1,18 +1,0 @@
-{
-  "Linuxspec.Spec.Namespaces.pid": {
-    "path": [
-      {
-        "/proc/1/ns/pid": "passed"
-      },
-      {
-        "proc/1/ns/ipc": "passed"
-      },
-      {
-        "proc/1/ns/uts": "passed"
-      },
-      {
-        "proc/1/ns/net": "passed"
-      }
-    ]
-  }
-}
