--- cases/specstest/report/platform_linux_amd64_out.json
@@ -1,10 +1,0 @@
-{
-  "Linuxspec.Spec.Platform": {
-    "OS": {
-      "linux": "passed"
-    },
-    "Arch": {
-      "amd64": "passed"
-    }
-  }
-}
