--- cases/specstest/report/readonly_true_out.json
@@ -1,5 +1,0 @@
-{
-  "Linuxspec.Spec.Root.Readonly": {
-    "true": "passed"
-  }
-}
