--- cases/specstest/report/version_correct_out.json
@@ -1,5 +1,0 @@
-{
-  "Linuxspec.Spec.Version": {
-    "pre-draft": "passed"
-  }
-}
