--- cases/specstest/source/config_newest.json
@@ -1,78 +1,0 @@
-{
-    "version": "pre-draft",
-    "platform": {
-        "os": "linux",
-        "arch": "amd64"
-    },
-    "process": {
-        "terminal": false,
-        "user": {
-            "uid": 0,
-            "gid": 0,
-            "additionalGids": null
-        },
-        "args": [
-	"/bin/ls"
-         ],
-        "env": null,
-        "cwd": ""
-    },
-    "root": {
-        "path": "",
-        "readonly": false
-    },
-    "hostname": "",
-    "mounts":null,
-    "hooks": {
-        "prestart": null,
-        "poststop": null
-    },
-    "linux": {
-        "uidMappings": null,
-        "gidMappings": null,
-        "rlimits": null,
-        "sysctl": null,
-        "resources": {
-            "disableOOMKiller": false,
-            "memory": {
-                "limit": 0,
-                "reservation": 0,
-                "swap": 0,
-                "kernel": 0,
-                "swappiness": -1
-            },
-            "cpu": {
-                "shares": 0,
-                "quota": 0,
-                "period": 0,
-                "realtimeRuntime": 0,
-                "realtimePeriod": 0,
-                "cpus": "",
-                "mems": ""
-            },
-            "blockIO": {
-                "blkioWeight": 0,
-                "blkioWeightDevice": "",
-                "blkioThrottleReadBpsDevice": "",
-                "blkioThrottleWriteBpsDevice": "",
-                "blkioThrottleReadIopsDevice": "",
-                "blkioThrottleWriteIopsDevice": ""
-            },
-            "hugepageLimits": null,
-            "network": {
-                "classId": "",
-                "priorities": null
-            }
-        },
-        "namespaces":null,
-        "capabilities": null,
-        "devices": null,
-        "apparmorProfile": "",
-        "selinuxProcessLabel": "",
-        "seccomp": {
-            "defaultAction": "",
-            "syscalls": null
-        },
-        "rootfsPropagation": ""
-    }
-}
