--- cases/specstest/source/specs/version.go
@@ -1,4 +1,0 @@
-package specs
-
-// Version is the specification version that the package types support.
-const Version = "pre-draft"
