--- cases/specstest/spec_process/source/process_guest.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"fmt"
-)
-
-func main() {
-	fmt.Println("runc ran!")
-}
