--- cases/specstest/spec_version/Specstest-spec_version-runc.json
@@ -1,57 +1,0 @@
-{
-  "Name": "Specstest-version_correct-runc",
-  "Summary": "test based opencontainers/specs",
-  "Owner": "linzhinan@huawei.com",
-  "Description": "Test runc when spec version == pre-draft",
-  "Group": "Specstest/version_correct/",
-  "License": "Apache 2.0",
-  "Explains": "Test runc when spec version == pre-draft",
-  "Requires": [
-    {
-      "Class": "OperationOS",
-      "Type": "os",
-      "Distribution": "ubuntu",
-      "Version": "14.04",
-      "Resource": {
-        "CPU": 1,
-        "Memory": "1GB",
-        "Disk": "2G"
-      }
-    },
-    {
-      "Class": "specstest",
-      "Type": "container",
-      "Distribution": "runc",
-      "Version": "0.2",
-      "Files": [
-        "./../../source/config.json"
-      ]
-    }
-  ],
-  "Deploys": [
-    {
-      "Object": "hostA",
-      "Class": "OperationOS",
-      "Files": [
-        "./source/version_host.go",
-        "./source/test_version_correct.go",
-        "./source/demo.go"
-      ],
-      "Cmd": "go build -o test_version ; ./test_version",
-      "Containers": [
-        {
-          "Object": "specs",
-          "Class": "specstest"
-        }
-      ]
-    }
-  ],
-  "Collects": [
-    {
-      "Object": "hostA",
-      "Files": [
-        "/tmp/testtool/spec_version.json"
-      ]
-    }
-  ]
-}
